Significant overhaul to TinyMCE robustness approach
Previously, a mix of targeted watching tools (both TinyMCE events and a DOMObserver) were used to detect when changes happened, and a relatively limited set of changes were expected.

TinyMCE can and will do anything, and the app now watches for *any* change inside the editor, and then fully re-scans the DOM both for new grids (in the case of copying and pasting a grid) as well as changed or deleted ones. Canvas Grid Builder can now also recover from detached nodes.

These likely come at a performance hit compared to the previous approach, but it is not expected to be significant in most circumstances. More testing needed.