Fix issue causing page hang on ImageCards
Previously, if an image card was destroyed unpredictably, the page would hang as the label inside the card would attempt to create and then self-delete itself. This happened because the element recognised that it was not in the editor window (deleted), but not that the parent it tried to attach to was also removed from the window.